+++ src/domain/mod.rs
@@ -18,5 +18,5 @@ pub use customers::customer::Entity as Customer;
 pub use lineitems::lineitem::Entity as Lineitem;
 pub use orders::order::Entity as Order;
 
-pub mod db;
-pub use db::{pool, migrations};
+// pub mod db;
+// pub use db::{pool, migrations};
